move service length constants to types
